pyproject.toml
[tool.poetry]
authors = ["hemengyang <hmy0119@gmail.com>"]
description = "基于 NoneBot 的聊天机器人"
license = "MIT"
name = "coolqbot"
readme = "README.md"
repository = "https://github.com/he0119/CoolQBot"
version = "0.14.0"

[tool.poetry.dependencies]
nonebot2 = {extras = ["full"], version = "^2.0.0-alpha.6"}
python = "~3.8"
python-dateutil = "^2.8.1"
gino = "^1.0.1"

[tool.poetry.dev-dependencies]
yapf = "*"

[[tool.poetry.source]]
default = true
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"

[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]
